requirements.txt
opencv-python
pyserial
pynput